requirements.txt
pandas
Pillow
python-pptx
openpyxl